Fix portability warnings in common Vulkan sources
    Add missing <cstdint> includes to resource_helpers.cxxm and
    destructors.cxx so the global-module fragment compiles without
    implicit declarations. Cast the frame_slot_index_ assignment and
    the VK_FENCE_CREATE_SIGNALED_BIT flag to silence -Wconversion.
    Parenthesize alignof(uint32_t) - 1 before the bitwise AND in the
    alignment check. Comment out the VK_ERROR_VALIDATION_FAILED case in
    the VkResult formatter to silence an enum-completeness warning on
    current SDK headers.

+++ src/common/vulkan/frame_ring.cxx
@@ -73,7 +73,7 @@ namespace vkgc
             return std::numeric_limits<std::uint32_t>::max();
         }
 
-        frame_slot_index_ = frame_index % kFramesInFlight_;
+        frame_slot_index_ = static_cast<std::uint32_t>(frame_index % kFramesInFlight_);
 
         {
             std::uint64_t const wait_value = frame_index - kFramesInFlight_;
